Implementation of SMO functions with empty call to other modules interface

+++ src/main/java/com/typhon/evolutiontool/services/EvolutionServiceImpl.java
@@ -1,9 +1,6 @@
 package com.typhon.evolutiontool.services;
 
-import com.typhon.evolutiontool.entities.Entity;
-import com.typhon.evolutiontool.entities.ParametersKeyString;
-import com.typhon.evolutiontool.entities.SMO;
-import com.typhon.evolutiontool.entities.WorkingSet;
+import com.typhon.evolutiontool.entities.*;
 import com.typhon.evolutiontool.exceptions.InputParameterException;
 import com.typhon.evolutiontool.services.typhonDL.TyphonDLInterface;
 import com.typhon.evolutiontool.services.typhonML.TyphonMLInterface;
@@ -14,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -72,6 +70,10 @@ public class EvolutionServiceImpl implements EvolutionService{
             entityname = smo.getInputParameter().get(ParametersKeyString.ENTITYNAME).toString();
             sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
             targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            //If the entity is involved in a relationship. Abort
+            if (typhonMLInterface.hasRelationship(entityname)) {
+                throw new InputParameterException("Cannot delete an entity involved in a relationship. Remove the relationships first.");
+            }
             //Delete data
             typhonInterface.deleteAllEntityData(entityname,sourcemodelid);
             //Delete structures
@@ -101,8 +103,8 @@ public class EvolutionServiceImpl implements EvolutionService{
 
 
     /**
-     * Migrates the instances data instances of sourceEntity that has a given attributeValue of their attribute
-     * attributeName to an existing targetEntity with the same attributes.
+     * Migrates the data instances of sourceEntity that has a given attributeValue of their attribute
+     * attributeName to an existing targetEntity with the same structure.
      * //TODO With a non already existing targetEntity
      * @param smo
      * @return
@@ -133,7 +135,7 @@ public class EvolutionServiceImpl implements EvolutionService{
             typhonInterface.deleteWorkingSetData(dataSource,sourcemodelid);
 
             typhonMLInterface.setNewTyphonMLModel(targetmodelid);
-            return "entity " + sourceEntityName + "splitted";
+            return "entity " + sourceEntityName + "split";
         } else {
             throw new InputParameterException("Missing parameter");
         }
@@ -149,8 +151,51 @@ public class EvolutionServiceImpl implements EvolutionService{
      */
     @Override
     public String splitVertical(SMO smo) throws InputParameterException {
+        String sourceEntityName, sourcemodelid, targetmodelid, databasetype, databasename, sourceEntityId;
+        Relation relation;
+        Entity targetEntity, sourceEntity;
+        List<String> attributes;
+        WorkingSet dataSource, dataTarget;
+        dataTarget = WorkingSetFactory.createEmptyWorkingSet();
+        if (containParameters(smo, Arrays.asList(
+                ParametersKeyString.ENTITY,
+                ParametersKeyString.TARGETMODEL,
+                ParametersKeyString.SOURCEMODEL,
+                ParametersKeyString.DATABASENAME,
+                ParametersKeyString.DATABASETYPE,
+                ParametersKeyString.ENTITYNAME,
+                ParametersKeyString.ATTRIBUTES))) {
+            sourceEntityName = smo.getInputParameter().get(ParametersKeyString.ENTITYNAME).toString();
+            targetEntity = smo.getPOJOFromInputParameter(ParametersKeyString.ENTITY, Entity.class);
+            attributes = smo.getPOJOFromInputParameter(ParametersKeyString.ATTRIBUTES, ArrayList.class);
+            targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
+            databasetype = smo.getInputParameter().get(ParametersKeyString.DATABASETYPE).toString();
+            databasename = smo.getInputParameter().get(ParametersKeyString.DATABASENAME).toString();
+            sourceEntity = typhonMLInterface.getEntityTypeFromId(sourceEntityName,sourcemodelid);
+            if (!typhonDLInterface.isDatabaseRunning(databasetype, databasename)) {
+                typhonDLInterface.createDatabase(databasetype, databasename);
+            }
+
+            relation = new Relation("splitVerticalResult", sourceEntity, targetEntity, null, false, Cardinality.ONE_ONE);
+            typhonInterface.createEntityType(targetEntity, targetmodelid);
+            this.createRelationshipType(relation, targetmodelid);
+            sourceEntityId = typhonMLInterface.getAttributeIdOfEntityType(sourceEntityName);
+            attributes.add(sourceEntityId);
+            dataSource = typhonInterface.readEntityDataSelectAttributes(sourceEntityName, attributes, sourcemodelid);
+            dataTarget.setEntityRows(targetEntity.getName(), dataSource.getEntityInstanceRows(sourceEntityName));
+            typhonInterface.writeWorkingSetData(dataTarget,targetmodelid);
+            attributes.remove(sourceEntityId);
+            typhonInterface.deleteAttributes(sourceEntityName, attributes, sourcemodelid);
+
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+
+            return "entity splitted vertically";
+        }
+
         return null;
     }
+
 
     /**
      * Migrates data of entity in sourceModel (read) to entity in targetModel (write).
@@ -164,8 +209,8 @@ public class EvolutionServiceImpl implements EvolutionService{
         Entity entity;
         String entityname, targetmodelid, databasetype, databasename, sourcemodelid;
         WorkingSet data;
-        if (containParameters(smo, Arrays.asList(ParametersKeyString.ENTITY, ParametersKeyString.TARGETMODEL, ParametersKeyString.SOURCEMODEL, ParametersKeyString.DATABASENAME, ParametersKeyString.DATABASETYPE))) {
-            entityname = smo.getInputParameter().get(ParametersKeyString.ENTITY).toString();
+        if (containParameters(smo, Arrays.asList(ParametersKeyString.ENTITYNAME, ParametersKeyString.TARGETMODEL, ParametersKeyString.SOURCEMODEL, ParametersKeyString.DATABASENAME, ParametersKeyString.DATABASETYPE))) {
+            entityname = smo.getInputParameter().get(ParametersKeyString.ENTITYNAME).toString();
             targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
             sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
             databasetype = smo.getInputParameter().get(ParametersKeyString.DATABASETYPE).toString();
@@ -179,6 +224,7 @@ public class EvolutionServiceImpl implements EvolutionService{
             data = typhonInterface.readAllEntityData(entity,sourcemodelid);
             typhonInterface.writeWorkingSetData(data,targetmodelid);
             typhonInterface.deleteWorkingSetData(data, sourcemodelid);
+            typhonInterface.deleteEntityStructure(entityname, sourcemodelid);
             typhonMLInterface.setNewTyphonMLModel(targetmodelid);
             return "entity migrated";
         } else {
@@ -193,32 +239,133 @@ public class EvolutionServiceImpl implements EvolutionService{
 
     @Override
     public String addRelationship(SMO smo) throws InputParameterException {
-        return null;
+        Relation relation;
+        String targetmodelid;
+        if (containParameters(smo, Arrays.asList(ParametersKeyString.RELATION, ParametersKeyString.TARGETMODEL))) {
+            relation = smo.getPOJOFromInputParameter(ParametersKeyString.RELATION, Relation.class);
+            targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            this.createRelationshipType(relation, targetmodelid);
+
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+            return "Relationship created";
+        } else {
+            throw new InputParameterException("Missing parameter");
+        }
     }
 
     @Override
     public String removeRelationship(SMO smo) {
+        boolean datadelete;
+        Relation relation;
+        String sourcemodelid, targetmodelid;
+        if (containParameters(smo, Arrays.asList(ParametersKeyString.SOURCEMODEL, ParametersKeyString.TARGETMODEL,ParametersKeyString.RELATION, ParametersKeyString.DATADELETE))) {
+            datadelete = Boolean.parseBoolean(smo.getInputParameter().get(ParametersKeyString.DATADELETE).toString());
+            relation = smo.getPOJOFromInputParameter(ParametersKeyString.RELATION, Relation.class);
+            sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
+            targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            typhonInterface.deleteForeignKey(relation.getSourceEntity(), relation.getTargetEntity());
+            if(datadelete)
+                typhonInterface.deleteAttributes(relation.getSourceEntity().getName(),Arrays.asList(typhonMLInterface.getAttributeOfType(relation.getSourceEntity().getName(),relation.getTargetEntity())),sourcemodelid);
+
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+            return "relationship type deleted";
+        }
+
         return null;
     }
 
     @Override
-    public String enableContainmentInRelationship(SMO smo) {
+    public String enableContainmentInRelationship(SMO smo) throws InputParameterException {
+        Relation relation;
+        String sourcemodelid, targetmodelid;
+        WorkingSet ws;
+        if (containParameters(smo, Arrays.asList(ParametersKeyString.SOURCEMODEL, ParametersKeyString.TARGETMODEL,ParametersKeyString.RELATION, ParametersKeyString.DATADELETE))) {
+            relation = smo.getPOJOFromInputParameter(ParametersKeyString.RELATION,Relation.class);
+            sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
+            targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            if (typhonMLInterface.getDatabaseType(relation.getSourceEntity().getName()) instanceof RelationalDB) {
+                throw new InputParameterException("Cannot produce a containment relationship in relational database source entity");
+            }
+            ws = typhonInterface.readRelationship(relation,sourcemodelid);
+            typhonInterface.writeWorkingSetData(ws, targetmodelid);
+            typhonInterface.deleteRelationship(relation, true, sourcemodelid);
+            // = delete Entity if relational. TODO
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+            return "Relationship containement enabled";
+        }
         return null;
     }
 
     @Override
-    public String disableContainmentInRelationship(SMO smo) {
+    public String disableContainmentInRelationship(SMO smo) throws InputParameterException {
+        Relation relation;
+        String sourcemodelid, targetmodelid;
+        WorkingSet ws;
+        if (containParameters(smo, Arrays.asList(ParametersKeyString.SOURCEMODEL, ParametersKeyString.TARGETMODEL,ParametersKeyString.RELATION))) {
+            relation = smo.getPOJOFromInputParameter(ParametersKeyString.RELATION,Relation.class);
+            sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
+            targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            if (typhonMLInterface.getDatabaseType(relation.getSourceEntity().getName()) instanceof RelationalDB) {
+                throw new InputParameterException("Please use splitHorizontal operation in case of relational model");
+            }
+            ws = typhonInterface.readRelationship(relation,sourcemodelid);
+            typhonInterface.createEntityType(relation.getTargetEntity(),targetmodelid);
+            typhonInterface.writeWorkingSetData(ws, targetmodelid);
+            typhonInterface.deleteRelationship(relation, true, sourcemodelid);
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+            return "Relationship containement disabled";
+        }
+        else{
+            throw new InputParameterException("Missing parameter");
+        }
+    }
+
+    @Override
+    public String enableOppositeRelationship(SMO smo) throws InputParameterException {
+        Relation relation, oppositeRel;
+        String sourcemodelid, targetmodelid, relationname;
+        if (containParameters(smo, Arrays.asList(ParametersKeyString.SOURCEMODEL, ParametersKeyString.TARGETMODEL,ParametersKeyString.RELATIONNAME))) {
+            sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
+            targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            relationname = smo.getInputParameter().get(ParametersKeyString.RELATIONNAME).toString();
+            relation = typhonMLInterface.getRelationFromName(relationname);
+            if (relation == null) {
+                throw new InputParameterException("No existing relationship with name provided relationship name");
+            }
+            //Quid opposite cardinality? TODO
+            oppositeRel = new Relation("opposite - " + relation.getName(), relation.getTargetEntity(), relation.getSourceEntity(), relation, false, relation.getCardinality());
+            this.createRelationshipType(oppositeRel, targetmodelid);
+
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+        }else{
+            throw new InputParameterException("Missing parameters");
+        }
         return null;
     }
 
     @Override
-    public String enableOppositeRelationship(SMO smo) {
-        return null;
-    }
-
-    @Override
-    public String disableOppositeRelationship(SMO smo) {
-        return null;
+    public String disableOppositeRelationship(SMO smo) throws InputParameterException {
+        Relation relation, oppositeRel;
+        String relationname, sourcemodelid, targetmodelid;
+        boolean datadelete;
+        if (containParameters(smo, Arrays.asList(ParametersKeyString.SOURCEMODEL, ParametersKeyString.TARGETMODEL,ParametersKeyString.RELATIONNAME, ParametersKeyString.DATADELETE))) {
+            sourcemodelid = smo.getInputParameter().get(ParametersKeyString.SOURCEMODEL).toString();
+            targetmodelid = smo.getInputParameter().get(ParametersKeyString.TARGETMODEL).toString();
+            relationname = smo.getInputParameter().get(ParametersKeyString.RELATIONNAME).toString();
+            datadelete = Boolean.parseBoolean(smo.getInputParameter().get(ParametersKeyString.DATADELETE).toString());
+            relation = typhonMLInterface.getRelationFromName(relationname);
+            if (relation == null) {
+                throw new InputParameterException("No existing relationship with name provided relationship name");
+            }
+            oppositeRel = relation.getOpposite();
+            typhonInterface.deleteRelationship(oppositeRel, datadelete, sourcemodelid);
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+            typhonMLInterface.setNewTyphonMLModel(targetmodelid);
+            return "opposite relationship deleted";
+        }
+        else{
+            throw new InputParameterException("Missing parameter");
+        }
     }
 
     @Override
@@ -304,6 +451,33 @@ public class EvolutionServiceImpl implements EvolutionService{
     public boolean containParameters(SMO smo, List<String> parameters) {
         logger.info("Verifying input parameter for [{}] - [{}] operator",smo.getTyphonObject(), smo.getEvolutionOperator());
         return smo.inputParametersContainsExpected(parameters);
+    }
+
+    private void createRelationshipType(Relation relation, String targetmodelid) {
+        // Implement here rules detailed in appendix file about actions on specific datamodels.
+
+        //If source & target are on relational
+        if(typhonMLInterface.getDatabaseType(relation.getSourceEntity().getName()) instanceof RelationalDB &&
+                typhonMLInterface.getDatabaseType(relation.getTargetEntity().getName()) instanceof RelationalDB)
+            switch (relation.getCardinality()) {
+                case N_N:
+                    typhonInterface.createJoinTable(relation.getSourceEntity(), relation.getTargetEntity());
+                    break;
+                case ONE_N:
+                    typhonInterface.addForeignKey(relation.getTargetEntity(), relation.getSourceEntity(),targetmodelid, true, false);
+                    break;
+                case ZERO_ONE:
+                    typhonInterface.addForeignKey(relation.getSourceEntity(), relation.getTargetEntity(),targetmodelid, false, true);
+                    break;
+                case ONE_ONE:
+                    typhonInterface.addForeignKey(relation.getSourceEntity(), relation.getTargetEntity(),targetmodelid, true, true);
+                    //+ data verification rule (all ids must referenced as fk).
+                    break;
+                case ZERO_N:
+                    typhonInterface.addForeignKey(relation.getTargetEntity(), relation.getSourceEntity(),targetmodelid, false, false);
+            } else{
+            // No specific action, but changes the way data is inserted. (Construction of key value pairs, or adding of reference attribute data.
+        }
     }
 
     public void setTyphonDLInterface(TyphonDLInterface typhonDLInterface) {
